third_party/WebKit/LayoutTests/fast/dom/Document/title-with-multiple-children-expected.txt

This test checks to see if setting document.title works even if the title element has multiple children.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.getElementsByTagName('title').length is 0
PASS document.title is ''
PASS titleElement.text is ''
PASS document.title is firstText
PASS titleElement.text is firstText
PASS document.title is firstText + secondText
PASS titleElement.text is firstText + secondText
PASS document.title is expected
PASS titleElement.text is expected
PASS successfullyParsed is true

TEST COMPLETE
